feat: add google oauth

+++ src/main/java/com/klikkas/dto/user_token/TokenUsages.java
@@ -7,6 +7,6 @@ public record TokenUsages(
     String llm_model,
     Integer input_token,
     Integer output_token,
-    Integer estimated_cost,
+    Double estimated_cost,
     LocalDateTime created_at
 ) {}
